# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FSP Talent API",
  "description": "\n**FSP Talent** — платформа подбора ИТ-специалистов с обратной механикой:\nработодатель сам находит кандидата в категоризированной базе и отправляет приглашение с вилкой ЗП.\n\n### Как авторизоваться в Swagger\n1. `POST /api/auth/register` → `POST /api/auth/verify-email` (в dev-режиме токен приходит в ответе регистрации).\n2. Кнопка **Authorize** вверху: username = email, password = пароль.\n\n### Формат ошибок\n`{\"code\": \"MACHINE_CODE\", \"message\": \"текст для пользователя\", \"detail\": \"...\"}`, для 422 ещё `fields: {поле: ошибка}`\n",
  "version": "1.3.0"
 },
 "paths": {
...
  "/api/employer/profile": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Профиль компании",
    "operationId": "get_profile_api_employer_profile_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EmployerProfileOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Обновить профиль компании",
    "operationId": "update_profile_api_employer_profile_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EmployerProfileUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EmployerProfileOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/employer/needs": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Мои сохранённые потребности («кого ищем»)",
    "operationId": "list_needs_api_employer_needs_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/NeedOut"
         },
         "type": "array",
         "title": "Response List Needs Api Employer Needs Get"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Сохранить новую потребность",
    "operationId": "create_need_api_employer_needs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NeedIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NeedOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/employer/needs/{need_id}": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Одна потребность",
    "operationId": "get_need_api_employer_needs__need_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "need_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Need Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NeedOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Изменить потребность (присылается целиком)",
    "operationId": "update_need_api_employer_needs__need_id__put",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "need_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Need Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NeedIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NeedOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Удалить потребность",
    "operationId": "delete_need_api_employer_needs__need_id__delete",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "need_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Need Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/candidates": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Подборка через параметры адреса (удобно для фильтров на фронте)",
    "description": "То же, что POST /candidates/search, но фильтры — в адресе. Пустая выдача: `items: [], total: 0`.",
    "operationId": "candidates_api_candidates_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "needs_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Взять критерии из сохранённой потребности",
       "title": "Needs Id"
      },
      "description": "Взять критерии из сохранённой потребности"
     },
     {
      "name": "specialization",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "backend",
          "frontend",
          "data_science",
          "qa",
          "devops"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Specialization"
      }
     },
     {
      "name": "grade",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "enum": [
         "intern",
         "junior",
         "middle",
         "senior"
        ],
        "type": "string"
       },
       "description": "Можно несколько: ?grade=junior&grade=middle",
       "default": [],
       "title": "Grade"
      },
      "description": "Можно несколько: ?grade=junior&grade=middle"
     },
     {
      "name": "stack",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "?stack=Python,Docker или ?stack=Python&stack=Docker",
       "default": [],
       "title": "Stack"
      },
      "description": "?stack=Python,Docker или ?stack=Python&stack=Docker"
     },
     {
      "name": "only_fsp",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Только с достижениями ФСП",
       "default": false,
       "title": "Only Fsp"
      },
      "description": "Только с достижениями ФСП"
     },
     {
      "name": "only_verified",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "Только с грейдом, подтверждённым тестом",
       "default": false,
       "title": "Only Verified"
      },
      "description": "Только с грейдом, подтверждённым тестом"
     },
     {
      "name": "min_test_score",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "maximum": 100,
         "minimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Test Score"
      }
     },
     {
      "name": "text",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 2000
        },
        {
         "type": "null"
        }
       ],
       "description": "Описание потребности свободным текстом",
       "title": "Text"
      },
      "description": "Описание потребности свободным текстом"
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/candidates/categories": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Обзор банка: категории и сколько в них кандидатов",
    "operationId": "categories_api_candidates_categories_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/CategoryBucket"
         },
         "type": "array",
         "title": "Response Categories Api Candidates Categories Get"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/candidates/search": {
   "post": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Подборка: фильтры + ранжирование + обоснование каждого кандидата",
    "description": "Выдача строится по КАТЕГОРИЯМ, присвоенным тестом (а не по тексту резюме).\nВ каждой карточке: `match_score` (0..100), `reasons` (плашки «почему он»), `breakdown` (вклад компонент).\nКонтакты (`contacts`) = null, пока кандидат не принял приглашение или не откликнулся сам.",
    "operationId": "search_api_candidates_search_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SearchIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SearchOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/candidates/{candidate_id}": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Карточка кандидата (контакты — только после принятия инвайта/отклика)",
    "operationId": "candidate_api_candidates__candidate_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "candidate_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Candidate Id"
      }
     },
     {
      "name": "vacancy_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Посчитать соответствие этой вакансии",
       "title": "Vacancy Id"
      },
      "description": "Посчитать соответствие этой вакансии"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CandidateDetail"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/shortlist": {
   "get": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Избранные кандидаты",
    "operationId": "shortlist_api_shortlist_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ShortlistOut"
         },
         "type": "array",
         "title": "Response Shortlist Api Shortlist Get"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Добавить в избранное",
    "operationId": "add_shortlist_api_shortlist_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ShortlistIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ShortlistOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт с текущим состоянием",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/shortlist/{candidate_id}": {
   "delete": {
    "tags": [
     "Работодатель: подбор"
    ],
    "summary": "Убрать из избранного",
    "operationId": "remove_shortlist_api_shortlist__candidate_id__delete",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "candidate_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Candidate Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/employer/integrations/ats/test": {
   "post": {
    "tags": [
     "Работодатель: интеграция с ATS"
    ],
    "summary": "Отправить в ATS тестовое событие",
    "operationId": "ats_test_api_employer_integrations_ats_test_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AtsDeliveryOut"
        }
       }
      }
     },
     "401": {
      "description": "Нет или неверный токен",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ErrorResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "items": {},
        "type": "array"
       }
      ],
      "title": "Detail"
     },
     "fields": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "string"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fields"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message",
     "detail"
    ],
    "title": "ErrorResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ShortlistIn": {
    "properties": {
     "candidate_id": {
      "type": "integer",
      "title": "Candidate Id"
     },
     "vacancy_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vacancy Id"
     },
     "note": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "candidate_id"
    ],
    "title": "ShortlistIn"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/token"
     }
    }
   }
  }
 }
}